Refactoring

+++ Engine/Player.hpp
@@ -9,8 +9,8 @@ public:
   Map::Vertex dir = {1, 0};
   Map::Vertex plane = {0, -1};
 
-  int realHeight = 50;
-  int height = 28;
+  double realHeight = 50;
+  double height = 28;
   double shakeState = 0;
 
   void rotate(double angle) {
